requirements.txt
# Python Dependencies for Backend

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# API Wrapper for Sleeper
sleeper-api-wrapper==1.0.6

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Date/Time Utilities
python-dateutil==2.8.2

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0


